Add files via upload
PongUPDATED scene is in folder Scenes -> Pong MiniGame Updated folder. 
Updated Pong MiniGame with an updated UI and *Working Score System. All it needs left is adding a cap to the score (best of 3 rounds?) because the score (and therefore the game...) goes on indefinitely until you manually stop the game.